BREAKING CHANGE - Modularizando a classe RelatorioSintético part relatorio de Pedidos
