+++ src/main/java/br/com/alura/comex/Main.java
@@ -1,6 +1,8 @@
 package br.com.alura.comex;
 
 
+import java.math.RoundingMode;
+import java.text.NumberFormat;
 import java.util.*;
 
 
@@ -11,17 +13,20 @@ public class Main {
         ProcessadorDeCsv arquivo = new ProcessadorDeCsv();
         List<Pedido> pedidos = arquivo.getPedidos("pedidos.csv");
 
-        GerarRelatorioPedidos relatorio = new GerarRelatorioPedidos(pedidos);
-        relatorio.imprimeValoresTotais();
-
         RelatorioClienteFieis relatorioClienteFieis = new RelatorioClienteFieis(pedidos);
         relatorioClienteFieis.printClientesFieis();
 
         RelatorioCategoria relatorioCategoria = new RelatorioCategoria(pedidos);
         relatorioCategoria.imprimeVendasPorCategoria();
 
-        relatorio.imprimeProdutosMaisVendidos();
+        RelatórioProdutos relatorioProduto = new RelatórioProdutos(pedidos);
+        relatorioProduto.imprimeProdutosMaisVendidos();
 
         relatorioCategoria.imprimeProdutoMaisCaroPorCategoria();
+
+        RelatorioSintetico relatorioSintetico = new RelatorioSintetico(pedidos);
+        relatorioSintetico.imprimeValoresTotais();
+
+
     }
 }
