+++ src/main/java/br/com/alura/comex/RelatorioSintetico.java
@@ -1,124 +1,35 @@
 package br.com.alura.comex;
 
 import java.math.BigDecimal;
+import java.math.RoundingMode;
+import java.text.NumberFormat;
 import java.util.*;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 public class RelatorioSintetico {
-
-    private int totalDeProdutosVendidos;
-    private int totalDePedidosRealizados;
-    private long totalDeCategorias;
-    private BigDecimal montanteDeVendas;
-    private Pedido pedidoMaisBarato;
-    private Pedido pedidoMaisCaro;
-    private Map<String, Integer> qtdProdutosPorCategoria;
-    private Map<String, BigDecimal> montantePorCategoria;
-    private Map<String, Integer> produtosMaisVendidos;
-    private Map<String, String> produtoMaisCaroCategoria;
-    private Map<String, BigDecimal> maiorPrecoPorCategoria;
+    private final List<Pedido> listaDePedidos;
 
     public RelatorioSintetico(List<Pedido> listaDePedidos){
 
         if (listaDePedidos == null || listaDePedidos.isEmpty()) throw new IllegalArgumentException("Lista de pedidos nula ou vazia!");
 
-        this.pedidoMaisBarato = listaDePedidos.stream()
-                .min(Comparator.comparing(Pedido::getValorTotal))
-                .orElseThrow(() -> new IllegalStateException("Lista está vazia"));
-
-        this.pedidoMaisCaro = listaDePedidos.stream()
-                .max(Comparator.comparing(Pedido::getValorTotal))
-                .orElseThrow(() -> new IllegalStateException("Lista está vazia"));
-
-        this.montanteDeVendas = listaDePedidos.stream()
-                .map(Pedido::getValorTotal)
-                .reduce(BigDecimal.ZERO, BigDecimal::add);
-
-        this.totalDeProdutosVendidos = listaDePedidos.stream()
-                .mapToInt(Pedido::getQuantidade)
-                .sum();
-
-        this.totalDePedidosRealizados = listaDePedidos.size();
-
-        this.totalDeCategorias = listaDePedidos.stream()
-                .map(Pedido::getCategoria)
-                .distinct()
-                .count();
-
-
-        this.qtdProdutosPorCategoria = new TreeMap<>();
-
-        listaDePedidos.stream()
-                .collect(Collectors.groupingBy(Pedido::getCategoria))
-                .forEach((a,v) -> qtdProdutosPorCategoria.put(a,v.stream().mapToInt(Pedido::getQuantidade).sum()));
-
-        this.montantePorCategoria = new TreeMap<>();
-        listaDePedidos.stream()
-                .collect(Collectors.groupingBy(Pedido::getCategoria))
-                .forEach((a,v) -> montantePorCategoria.put(a, v.stream().map(Pedido::getValorTotal).reduce(BigDecimal.ZERO, BigDecimal::add)));
-
-        this.produtosMaisVendidos = new TreeMap<>();
-        listaDePedidos.stream()
-                .collect(Collectors.groupingBy(Pedido::getProduto))
-                .forEach((a,v) -> produtosMaisVendidos.put(a,v.stream().mapToInt(Pedido::getQuantidade).sum()));
-
-        /* Essa solução não funciona se a fonte fosse um banco de dados, por exemplo */
-        this.produtoMaisCaroCategoria = new TreeMap<>();
-        listaDePedidos.stream().collect(Collectors.groupingBy(Pedido::getCategoria)).forEach((a, b) -> {
-            produtoMaisCaroCategoria.put(a, b.stream().max(Comparator.comparing(Pedido::getPreco))
-                    .orElseThrow(() -> new IllegalStateException("Não foi possível encontrar o produto mais caro da Categoria: " + a)).getProduto());
-        });
-
-        this.maiorPrecoPorCategoria = new TreeMap<>();
-        listaDePedidos.stream().collect(Collectors.groupingBy(Pedido::getCategoria)).forEach((a, b) -> {
-            maiorPrecoPorCategoria.put(a, b.stream().max(Comparator.comparing(Pedido::getPreco))
-                    .orElseThrow(() -> new IllegalStateException("Não foi possível encontrar o maior valor Categoria: " + a)).getPreco());
-        });
-
+        this.listaDePedidos = listaDePedidos;
     }
 
-    public int getTotalDeProdutosVendidos() {
-        return totalDeProdutosVendidos;
+    public void imprimeValoresTotais(){
+
+        RelatorioPedidos relatorioPedidos = new RelatorioPedidos(listaDePedidos);
+        RelatórioProdutos relatorioProduto = new RelatórioProdutos(listaDePedidos);
+        RelatorioCategoria relatorioCategoria = new RelatorioCategoria(listaDePedidos);
+
+        System.out.println("#### RELATÓRIO DE VALORES TOTAIS");
+        System.out.printf("- TOTAL DE PEDIDOS REALIZADOS: %s\n", relatorioPedidos.getTotalDePedidosRealizados());
+        System.out.printf("- TOTAL DE PRODUTOS VENDIDOS: %s\n", relatorioProduto.getTotalDeProdutosVendidos());
+        System.out.printf("- TOTAL DE CATEGORIAS: %s\n", relatorioCategoria.getTotalDeCategorias());
+        System.out.printf("- MONTANTE DE VENDAS: %s\n", NumberFormat.getCurrencyInstance(new Locale("pt", "BR")).format(relatorioPedidos.getMontanteDeVendas().setScale(2, RoundingMode.HALF_DOWN)));
+        System.out.printf("- PEDIDO MAIS BARATO: %s (%s)\n", NumberFormat.getCurrencyInstance(new Locale("pt", "BR")).format(relatorioPedidos.getPedidoMaisBarato().getValorTotal().setScale(2, RoundingMode.HALF_DOWN)), relatorioPedidos.getPedidoMaisBarato().getProduto());
+        System.out.printf("- PEDIDO MAIS CARO: %s (%s)\n", NumberFormat.getCurrencyInstance(new Locale("pt", "BR")).format(relatorioPedidos.getPedidoMaisCaro().getValorTotal().setScale(2, RoundingMode.HALF_DOWN)), relatorioPedidos.getPedidoMaisCaro().getProduto());
     }
 
-    public int getTotalDePedidosRealizados() {
-        return totalDePedidosRealizados;
-    }
-
-    public BigDecimal getMontanteDeVendas() {
-        return montanteDeVendas;
-    }
-
-    public Pedido getPedidoMaisBarato() {
-        return pedidoMaisBarato;
-    }
-
-    public Pedido getPedidoMaisCaro() {
-        return pedidoMaisCaro;
-    }
-
-    public long getTotalDeCategorias() {
-        return totalDeCategorias;
-    }
-
-    public Map<String, Integer> getQtdProdutosPorCategoria() {
-        return qtdProdutosPorCategoria;
-    }
-
-    public Map<String, BigDecimal> getMontantePorCategoria() {
-        return montantePorCategoria;
-    }
-
-    public Map<String, Integer> getProdutosMaisVendidos() {
-        return produtosMaisVendidos;
-    }
-
-    public Map<String, String> getProdutoMaisCaroCategoria() {
-        return produtoMaisCaroCategoria;
-    }
-
-    public Map<String, BigDecimal> getMaiorPrecoPorCategoria() {
-        return maiorPrecoPorCategoria;
-    }
 }
